OCSADV-200: move ProgramIdGen to spModel.core bundle
